Add files via upload
I recently completed a master's degree in finance and I was very interested in the topic of financial assets as well as underlying assets. Options are an underlying asset that behaves similarly to the Brownian motion of particles (physics), volatility, trend, and other variables interfere with the prediction of future prices. The following program shows the B&Ch method in open source for you to use in your calculations and models. This code is free and there should be no charge for its use, its focus is purely educational.